requirements.txt
astroid==2.3.3
autopep8==1.4.4
Django==2.2.7
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
